+++ R/shapviz.R
@@ -94,17 +94,16 @@ shapviz.matrix = function(object, X, baseline = 0, collapse = NULL, ...) {
 #'
 #' # Will use numeric matrix "X_pred" as feature matrix
 #' x <- shapviz(fit, X_pred = X_pred)
-#' x
-#' sv_importance(x)
+#' sv_importance(x, kind = "bar")
 #'
 #' # Will use original values as feature matrix
 #' x <- shapviz(fit, X_pred = X_pred, X = iris[, -1])
-#' x
 #' sv_dependence(x, "Petal.Length", color_var = "auto")
 #'
 #' # "X_pred" can also be passed as xgb.DMatrix, but only if X is passed as well!
-#' shapviz(fit, X_pred = dtrain, X = iris[, -1])
+#' x <- shapviz(fit, X_pred = dtrain, X = iris[, -1])
 #'
+#' # Similarly with LightGBM
 #' if (requireNamespace("lightgbm", quietly = TRUE)) {
 #'   fit <- lightgbm::lgb.train(
 #'     params = list(objective = "regression"),
@@ -113,18 +112,14 @@ shapviz.matrix = function(object, X, baseline = 0, collapse = NULL, ...) {
 #'     verbose = -2
 #'   )
 #'   x <- shapviz(fit, X_pred = X_pred)
-#'   sv_force(x, row_id = 1)
 #' }
 #'
 #' # In multiclass setting, we need to specify which_class (integer starting at 1)
-#' # should be explained.
-#' params <- list(objective="multi:softprob", num_class = 3)
+#' params <- list(objective = "multi:softprob", num_class = 3)
 #' X_pred <- data.matrix(iris[, -5])
 #' dtrain <- xgboost::xgb.DMatrix(X_pred, label = as.integer(iris[, 5]) - 1L)
 #' fit <- xgboost::xgb.train(params = params, data = dtrain, nrounds = 50)
 #' x <- shapviz(fit, X_pred = X_pred, which_class = 3)
-#' x
-#' sv_waterfall(x, row_id = 1)
 #'
 #' # What if we would have one-hot-encoded values and want to explain the original column?
 #' X_pred <- stats::model.matrix(~ . -1, iris[, -1])
@@ -136,8 +131,6 @@ shapviz.matrix = function(object, X, baseline = 0, collapse = NULL, ...) {
 #'   X = iris[, -1],
 #'   collapse = list(Species = c("Speciessetosa", "Speciesversicolor", "Speciesvirginica"))
 #' )
-#' x
-#' sv_force(x, row_id = 1)
 shapviz.xgb.Booster = function(object, X_pred, X = X_pred,
                                which_class = NULL, collapse = NULL, ...) {
   stopifnot(
@@ -178,10 +171,14 @@ shapviz.lgb.Booster = function(object, X_pred, X = X_pred,
     "X_pred must be a matrix" = is.matrix(X_pred),
     "X_pred must have column names" = !is.null(colnames(X_pred))
   )
-  if (utils::packageVersion("lightgbm") >= 4) {
-    S <- stats::predict(object, data = X_pred, type = "contrib", ...)
+
+  # Switch for different versions of predict.lgb.Booster()
+  is_v4 <- utils::packageVersion("lightgbm") >= 4
+  has_type <- "type" %in% names(formals(utils::getS3method("predict", "lgb.Booster")))
+  if (is_v4 || has_type) {
+    S <- stats::predict(object, newdata = X_pred, type = "contrib", ...)
   } else {
-    S <- stats::predict(object, data = X_pred, predcontrib = TRUE, ...)
+    S <- stats::predict(object, X_pred, predcontrib = TRUE, ...)
   }
 
   pp <- ncol(X_pred) + 1L