+++ NEWS.md
@@ -14,6 +14,7 @@
 - Added argument `viridis_args = getOption("shapviz.viridis_args")` to `sv_dependence()` and `sv_importance()` to control the viridis color scale options. The default global option equals `list(begin = 0.25, end = 0.85, option = "inferno")`. For example, to switch to a standard viridis scale, you can either change the default with `options(shapviz.viridis_args = NULL)` 
 or set `viridis_args = NULL`.
 - Deprecated helper functions `shapviz_from_lgb_predict()` and `shapviz_from_xgb_predict` in favour of the collapsing logic (see above). The functions will be removed in version 0.3.0.
+- Added 'lightgbm' as "Enhances" dependency.
 - More unit tests.
 - Improved documentation.
 - Added H2O section to vignette.