+++ NEWS.md
@@ -14,7 +14,9 @@
 
 - Added `dim()` method for "shapviz" object, implying `nrow()` and `ncol()`.
 - To allow more flexible formatting, the `format_fun` argument of `sv_waterfall()` and `sv_force()` has been replaced by `format_shap` to format SHAP values and `format_feat` to format numeric feature values. By default, they use the new global options "shapviz.format_shap" and "shapviz.format_feat", both with default `function(z) prettyNum(z, digits = 3, scientific = FALSE)`.
-- To be consistent with above change, the argument `format_fun` in `sv_importance()` was renamed to `format_shap` and the default also points to the global option `shapviz.format_shap`.
+- The argument `format_fun` in `sv_importance()` now uses a better aligned number formatter.
+- `sv_importance()` received an argument `show_numbers` to add or hide SHAP feature importance values.
+- Removed y-axis ticks in `sv_importance()`.
 - `sv_waterfall()` now uses the more consistent argument `order_fun = function(s) order(abs(s))` instead of the original `sort_fun = function(shap) abs(shap)` that was then passed to `order()`.
 - Added argument `viridis_args = getOption("shapviz.viridis_args")` to `sv_dependence()` and `sv_importance()` to control the viridis color scale options. The default global option equals `list(begin = 0.25, end = 0.85, option = "inferno")`. For example, to switch to a standard viridis scale, you can either change the default with `options(shapviz.viridis_args = NULL)` 
 or set `viridis_args = NULL`.